+++ Example/EpoxyExample/ViewControllers/LayoutGroups/LayoutGroupsReadmeExamplesViewController.swift
@@ -5,42 +5,36 @@ import EpoxyCollectionView
 import UIKit
 
 class LayoutGroupsReadmeExamplesViewController: CollectionViewController {
-
+  
   // MARK: Lifecycle
-
+  
   init() {
     super.init(layout: UICollectionViewCompositionalLayout.list)
     setItems(items, animated: false)
   }
-
+  
   // MARK: Internal
-
-  override var title: String? {
-    get { "Readme Examples" }
-    set { }
-  }
-
+  
+  @ItemModelBuilder
   var items: [ItemModeling] {
-    [
-      ActionButtonRow.itemModel(
-        dataID: DataID.actionButtonRow,
-        content: .init(
-          title: "Title text",
-          subtitle: "Subtitle text",
-          actionText: "Perform action")),
-      IconRow.itemModel(
-        dataID: DataID.iconRow,
-        content: .init(
-          title: "This is an IconRow",
-          icon: UIImage(systemName: "person.fill")!)),
-    ]
+    ActionButtonRow.itemModel(
+      dataID: DataID.actionButtonRow,
+      content: .init(
+        title: "Title text",
+        subtitle: "Subtitle text",
+        actionText: "Perform action"))
+    IconRow.itemModel(
+      dataID: DataID.iconRow,
+      content: .init(
+        title: "This is an IconRow",
+        icon: UIImage(systemName: "person.fill")!))
   }
-
+  
   // MARK: Private
-
+  
   private enum DataID {
     case actionButtonRow
     case iconRow
   }
-
+  
 }
